Fix: Release 빌드 경고 수정 (C4244 int/float/double 변환)
- ScoreSystem.cpp: AdditionScore를 int로 변경, 콜백에 static_cast<float> 추가
- BeatHUDWidget.cpp: double 리터럴에 f 접미사 추가

+++ Gameplay/ScoreSystem.cpp
@@ -9,7 +9,7 @@ void FScoreSystem::Reset()
 
 void FScoreSystem::AddBeatBonus(EBeatJudge Judge)
 {
-    float AdditionScore = 0.f;
+    int AdditionScore = 0;
     switch (Judge)
     {
     case EBeatJudge::Perfect:
@@ -29,7 +29,7 @@ void FScoreSystem::AddBeatBonus(EBeatJudge Judge)
 
     if (OnJudgeCallback)
     {
-        OnJudgeCallback(Judge, AdditionScore, Combo);
+        OnJudgeCallback(Judge, static_cast<float>(AdditionScore), Combo);
     }
     if (OnJudgeScoreUpdateCallback)
     {
